Move tests to test folder, use public API
